requirements.txt
langchain-core==0.3.66
langchain-openai==0.3.25
langgraph==0.4.8
openai>=1.86.0
pydantic==2.11.7
typing-extensions==4.12.2
matplotlib==3.9.4
google-adk==1.4.2
litellm>=1.35.0
python-dotenv==1.0.0
pytz==2024.1
requests>=2.32.4
duckduckgo-search==6.3.5
jinja2==3.1.6
pyyaml==6.0.2

# Document reading dependencies (optional)
PyPDF2>=3.0.0
python-docx>=1.1.0
python-pptx>=0.6.21
pandas>=2.0.0
openpyxl>=3.1.0

# Serving the end point.
fastapi>=0.115.0
uvicorn>=0.34.0
pydantic>=2.7.4
